chore(app): resolve residual ESLint warnings + track convert-to-synced follow-ups (#436)
* docs(todo): track convert-to-synced polish follow-ups

Move the four follow-ups flagged across the convert-to-synced PR
cycle (#431–#435) into docs/TODO.md so they surface during a future
Phase 6 polish pass instead of getting lost in PR discussion history.

- Friendly error-message mapping (413 / AbortError / 5xx → human text).
- aria-live region so convert state announces for screen readers.
- Concurrency cap on parallel asset uploads.
- Tighten internal composeWithTimeout / default-helper signatures from
  optional to explicit `| undefined`.


* chore(app): resolve residual ESLint warnings

Four warnings had accumulated across earlier PRs; bringing the app
package to zero so new work stands out instead of blending in.

- react-refresh/only-export-components on AuthContext.tsx: the file
  exported both the AuthProvider component and the useAuth hook,
  which breaks fast refresh when the hook is edited. Split into
  AuthContext.tsx (provider + AuthContext consumption) and a new
  useAuth.ts (hook + context + value type). Matches the existing
  DocumentManagerContext / useDocumentManagerContext split. Updates
  the two importers (App.tsx, DocumentSidebar.tsx).

- react-hooks/exhaustive-deps on handleOnline in
  OfflineAwareSyncedContainer.tsx: currentSessionId was read inside
  the callback but missing from the deps array. Added.

- react-hooks/exhaustive-deps + "complex expression in dep array" on
  the offline-flush effect in OfflineAwareSyncedContainer.tsx: the
  dep array had `mode.type === "offline" ? mode.recovery : undefined`
  (static-analysis hostile) and missed `mode.recovery` directly.
  Extract to a top-level `offlineRecovery` variable, use it both in
  the effect body and as a flat dep, and drop the redundant
  `mode.type` dep and inner guard since `offlineRecovery` already
  captures the "only run when offline" condition.


---------
