+++ packages/app/src/auth/useAuth.ts
@@ -1,0 +1,20 @@
+import { createContext, useContext } from "react";
+import type { User } from "./types";
+
+export interface AuthContextValue {
+  user: User | null;
+  loading: boolean;
+  loginWithGitHub: () => void;
+  loginWithGoogle: () => void;
+  logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextValue | null>(null);
+
+export function useAuth(): AuthContextValue {
+  const ctx = useContext(AuthContext);
+  if (!ctx) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return ctx;
+}
